use library for qrCode generation
