+++ ui/app/components/mfa-setup-step-one.js
@@ -18,17 +18,38 @@ export default class MfaSetupStepOne extends Component {
   @service store;
   @tracked error;
   @tracked warning;
+  @tracked qrCode = '';
+
+  @action
+  redirectPreviousPage() {
+    this.args.restartFlow();
+  }
+
+  @action
+  async verifyUUID(evt) {
+    evt.preventDefault();
+    let response = await this.postAdminGenerate();
+    this.args.saveUUIDandQrCode(this.UUID, this.qrCode); // parent needs to keep track of UUID and qrCode.
+    if (response === 'stop_progress') {
+      this.args.isUUIDVerified(false);
+    } else if (response === 'reset_method') {
+      this.args.showWarning(this.warning);
+    } else {
+      this.args.isUUIDVerified(true);
+    }
+  }
 
   async postAdminGenerate() {
     this.error = null;
+    this.warning = null;
     let adapter = this.store.adapterFor('mfa-setup');
     let response;
     try {
       response = await adapter.adminGenerate({
         entity_id: this.args.entityId,
-        method_id: this.UUID,
+        method_id: this.UUID, // comes from value on the input
       });
-      // if there was a warning it won't fail but needs to be handled here
+      // if there was a warning it won't fail but needs to be handled here and the flow needs to be interrupted
       let warnings = response.warnings || [];
       if (warnings.length > 0) {
         this.UUID = ''; // clear UUID
@@ -37,36 +58,19 @@ export default class MfaSetupStepOne extends Component {
         );
         if (alreadyGenerated) {
           // replace warning because it comes in with extra quotes: "Entity already has a secret for MFA method ""' "
-          // ARG TODO confirm with Ivana on language
           this.warning = 'Entity already has a secret for MFA method';
           return 'reset_method';
         }
         this.warning = warnings; // in case other kinds of warnings comes through. Still push to third screen because it's not an error.
         return 'reset_method';
       }
+      // if no warning, then set the qrCode which comes from the successful response.
+      this.qrCode = response.data.url;
     } catch (error) {
       this.UUID = ''; // clear the UUID
       this.error = error.errors;
       return 'stop_progress';
     }
     return response;
-  }
-
-  @action
-  async verifyUUID(evt) {
-    evt.preventDefault();
-    this.args.saveUUID(this.UUID); // send UUID to the parent. Needs to record in case of reset method.
-    let response = await this.postAdminGenerate();
-    if (response === 'stop_progress') {
-      this.args.isUUIDVerified(false);
-    } else if (response === 'reset_method') {
-      this.args.goToReset(this.warning);
-    } else {
-      this.args.isUUIDVerified(true);
-    }
-  }
-  @action
-  redirectPreviousPage() {
-    window.history.back();
   }
 }
