--- ui/app/templates/components/mfa-setup-step-three.hbs
@@ -1,13 +1,0 @@
-<p>
-  TOTP Multi-factor authentication (MFA) can be enabled here if it is required by your administrator. This will ensure that
-  you are not prevented from logging into Vault in the future, once MFA is fully enforced.
-</p>
-{{#if @warning}}
-  <AlertBanner @type="info" @title={{@warning}} class="has-top-margin-l" />
-{{/if}}
-<button id="restart" type="button" class="button has-text-danger" {{on "click" this.restartSetup}}>
-  Restart setup
-</button>
-<button id="cancel" type="button" {{on "click" this.redirectPreviousPage}} class="button">
-  Cancel
-</button>
