+++ ui/package.json
@@ -124,6 +124,7 @@
     "ember-modifier": "^3.1.0",
     "ember-page-title": "^6.2.2",
     "ember-power-select": "^5.0.3",
+    "ember-qrcode-shim": "^0.4.0",
     "ember-qunit": "^5.1.5",
     "ember-resolver": "^8.0.3",
     "ember-responsive": "^3.0.0-beta.3",
